+++ CHANGELOG.md
@@ -126,6 +126,10 @@ compute_plan = client.add_compute_plan(
 - Register functions in substratools using decorator `@tools.register` (#310)
 - Update README image (#318)
 
+### Removed
+
+- BREAKING CHANGE: failed task category
+
 ## [0.39.0](https://github.com/Substra/substra/releases/tag/0.39.0) - 2022-10-03
 
 ### Removed